+++ UFCTowerDefense.js
@@ -405,6 +405,7 @@ Game_TDEnemy.prototype.initialize = function (enemyName, spawnId) {
   this.setDirection(this._direction);
   this._realMoveSpeed = +this._enemyData.moveSpeed;
   this._moveSpeed = this._realMoveSpeed;
+  this._moveSpeedEffects = {};
   this._isStun = false;
   this._animationPlaying = false;
   this._event = new PIXI.utils.EventEmitter();
@@ -459,17 +460,18 @@ Game_TDEnemy.prototype.updateEffects = function () {
     if (!this._effects[effect].effect) continue;
 
     if (!this._effects[effect].enable) {
+      if (!this._effects[effect].effect.getChanceEffect()) {
+        this._effects[effect].enable = false;
+        this._effects[effect].effect = null;
+        continue;
+      }
       switch (effect) {
         case TowerDefenseManager.EFFECTS.COLD:
-          // Always pick the most effective effect
-          this._moveSpeed = Math.min(
-            this._moveSpeed,
-            Math.max(
-              0.1,
-              (this._realMoveSpeed *
-                (100 - this._effects[effect].effect.getEffect())) /
-                100
-            )
+          this.addMoveSpeedEffect(
+            effect,
+            -this._realMoveSpeed *
+              (this._effects[effect].effect.getEffect() / 100),
+            true
           );
           break;
         case TowerDefenseManager.EFFECTS.POISON:
@@ -478,11 +480,15 @@ Game_TDEnemy.prototype.updateEffects = function () {
           );
           break;
         case TowerDefenseManager.EFFECTS.STUN:
-          if (!this._effects[effect].effect.getChanceEffect()) {
-            this._effects[effect].effect = null;
-            return;
-          }
           this._isStun = true;
+          break;
+        case TowerDefenseManager.EFFECTS.RAGE:
+          this.addMoveSpeedEffect(
+            effect,
+            this._realMoveSpeed *
+              (this._effects[effect].effect.getEffect() / 100),
+            false
+          );
           break;
       }
       this._event.emit("addEffect", effect);
@@ -492,7 +498,8 @@ Game_TDEnemy.prototype.updateEffects = function () {
     if (this._effects[effect].effect.isDone()) {
       switch (effect) {
         case TowerDefenseManager.EFFECTS.COLD:
-          this._moveSpeed = this._realMoveSpeed;
+        case TowerDefenseManager.EFFECTS.RAGE:
+          this.removeMoveSpeedEffect(effect);
           break;
         case TowerDefenseManager.EFFECTS.STUN:
           this._isStun = false;
@@ -503,6 +510,37 @@ Game_TDEnemy.prototype.updateEffects = function () {
       this._event.emit("removeEffect", effect);
     }
   }
+};
+
+Game_TDEnemy.prototype.updateMoveSpeed = function () {
+  if (Object.keys(this._moveSpeedEffects).length >= 0) {
+    let newSpeed = this._realMoveSpeed;
+    for (let speedEffect in this._moveSpeedEffects) {
+      newSpeed += this._moveSpeedEffects[speedEffect];
+    }
+    this._moveSpeed = newSpeed;
+  } else {
+    this._moveSpeed = this._realMoveSpeed;
+  }
+};
+
+Game_TDEnemy.prototype.removeMoveSpeedEffect = function (effect) {
+  delete this._moveSpeedEffects[effect];
+  this.updateMoveSpeed();
+};
+
+Game_TDEnemy.prototype.addMoveSpeedEffect = function (effect, value, isSlow) {
+  if (!this._moveSpeedEffects[effect]) {
+    this._moveSpeedEffects[effect] = value;
+  } else if (
+    (this._moveSpeedEffects[effect] < value && isSlow) ||
+    (this._moveSpeedEffects[effect] > value && !isSlow)
+  ) {
+    this._moveSpeedEffects[effect] = value;
+  } else {
+    return;
+  }
+  this.updateMoveSpeed();
 };
 
 Game_TDEnemy.prototype.distancePerFrame = function () {
@@ -969,6 +1007,9 @@ Sprite_ufcTDEnemy.prototype.addEffect = function (effect) {
       break;
     case TowerDefenseManager.EFFECTS.STUN:
       _effect = 5;
+      break;
+    case TowerDefenseManager.EFFECTS.RAGE:
+      _effect = 4;
       break;
   }
   this.addChild(new Sprite_ufcTDEnemyEffect(effect, _effect, 0, 25));
@@ -1948,6 +1989,7 @@ TowerDefenseManager.EFFECTS = {
   COLD: "cold",
   POISON: "poison",
   STUN: "stun",
+  RAGE: "rage",
 };
 
 TowerDefenseManager.setLimitAnimation = function (limit) {
@@ -2272,8 +2314,9 @@ ufcTowerData.prototype.initialize = function (data) {
       let _effect = effect.split("|");
       this._effects.push({
         name: _effect[0],
-        effect: _effect[1],
-        duration: _effect[2],
+        effect: +_effect[1],
+        duration: +_effect[2],
+        chance: +_effect[3] || 100,
       });
     }
   }
@@ -2301,7 +2344,8 @@ const ufcTowerEffects = function () {
 ufcTowerEffects.prototype.initialize = function (data) {
   this._name = data.name;
   this._effect = data.effect;
-  this._duration = +data.duration;
+  this._duration = data.duration;
+  this._chance = data.chance;
   this._curTime = this._duration;
   this._isDone = false;
   this._effectPerSecond = false;
@@ -2329,7 +2373,7 @@ ufcTowerEffects.prototype.getEffect = function () {
 };
 
 ufcTowerEffects.prototype.getChanceEffect = function () {
-  return this._effect > Math.randomInt(100);
+  return this._chance < 100 ? this._chance > Math.randomInt(100) : true;
 };
 
 ufcTowerEffects.prototype.isDone = function () {