+++ dev/towerdefense/js/Sprite_ufcTDEnemy.js
@@ -45,6 +45,9 @@ Sprite_ufcTDEnemy.prototype.addEffect = function (effect) {
     case TowerDefenseManager.EFFECTS.STUN:
       _effect = 5;
       break;
+    case TowerDefenseManager.EFFECTS.RAGE:
+      _effect = 4;
+      break;
   }
   this.addChild(new Sprite_ufcTDEnemyEffect(effect, _effect, 0, 25));
 };